Move Sudoku tests up a dir
